requirements.txt
beautifulsoup4==4.11.1
dnspython==2.2.1
geoip2==4.5.0
mysql-connector-python==8.0.29
pandas==1.3.4
python_whois==0.7.3
requests==2.26.0
scikit_learn==1.1.1
tldextract==3.2.0
urllib3==1.26.7
whois==0.9.16
seaborn==0.11.2
xgboost==1.6.1